First Prototype created
this is just a sketch for what is to come

+++ backend/src/jsons/check_request.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug,Deserialize, Serialize)]
+#[derive(Debug,Deserialize, Serialize,Default)]
 pub struct CheckRequest{
-    pub default : i32,
-    pub fallback: i32,
+    pub default : isize,
+    pub fallback: isize,
 }
